Reference

XCALibre.Discretise.Dirichlet — Type

Dirichlet <: AbstractDirichlet

Dirichlet boundary condition model.

Fields

'ID' – Boundary ID

value – Scalar or Vector value for Dirichlet boundary condition.

XCALibre.Discretise.DirichletFunction — Type

DirichletFunction <: AbstractBoundary

Dirichlet boundary condition model.

Fields

'ID' – Boundary ID

value – Custom function for Dirichlet boundary condition.

XCALibre.Discretise.FixedTemperature — Type

FixedTemperature <: AbstractDirichlet

Fixed temperature boundary condition model, which allows the user to specify wall temperature that can be translated to the energy specific model, such as sensivle enthalpy.

Fields

'ID' – Boundary ID

value – Scalar or Vector value for Dirichlet boundary condition.

T - Temperature value in Kelvin.

model - Energy physics model for case.

Examples

FixedTemperature(:inlet, T=300.0, model=model.energy),

XCALibre.Discretise.Neumann — Type

Neumann <: AbstractNeumann

Neumann boundary condition model (currently only configured for zero gradient)

Fields

'ID' – Boundary ID

value – Scalar or Vector value for Neumann boundary condition.

XCALibre.Discretise.Periodic — Type

Periodic <: AbstractBoundary

Periodic boundary condition model.

Fields

'ID' – Boundary ID

XCALibre.Discretise.Symmetry — Type

Symmetry <: AbstractNeumann

Symmetry boundary condition model for velocity and scalar field.

Fields

'ID' – Boundary ID

XCALibre.Discretise.Wall — Type

Wall <: AbstractDirichlet

Wall boundary condition model for no-slip wall condition.

Fields

'ID' – Boundary ID

XCALibre.Discretise.construct_periodic — Method

construct_periodic(mesh, backend, patch1::Symbol, patch2::Symbol)

Function for construction of periodic boundary conditions.

Input

mesh – Mesh.

backend – Backend configuraton.

patch1 – Primary periodic patch ID.

patch2 – Neighbour periodic patch ID.

Output

KOmegaLKEModel( k_eqn, ω_eqn, kl_eqn, nueffkLS, nueffkS, nueffωS, nuL, nuts, Ω, γ, fv, normU, Reυ, ∇k, ∇ω ) – Turbulence model structure.

Example

- `periodic = construct_periodic(mesh, CPU(), :top, :bottom)` - Example using CPU 
backend with periodic boundaries named top and bottom.

XCALibre.ModelPhysics.Compressible — Type

Compressible <: AbstractCompressible

Compressible fluid model containing fluid field parameters for compressible flows with constant parameters - ideal gas with constant viscosity.

Fields

'nu' – Fluid kinematic viscosity.

'cp' – Fluid specific heat capacity.

gamma – Ratio of specific heats.

Pr – Fluid Prantl number.

Examples

Fluid{Compressible}(; nu=1E-5, cp=1005.0, gamma=1.4, Pr=0.7) - Constructur with default values.

XCALibre.ModelPhysics.Fluid — Type

Fluid <: AbstractFluid

Abstract fluid model type for consturcting new fluid models.

Fields

'args' – Model arguments.

XCALibre.ModelPhysics.Incompressible — Type

Incompressible <: AbstractIncompressible

Incompressible fluid model containing fluid field parameters for incompressible flows.

Fields

'nu' – Fluid kinematic viscosity.

'rho' – Fluid density.

Examples

Fluid{Incompressible}(nu=0.001, rho=1.0) - Constructor with default values.

XCALibre.ModelPhysics.KOmega — Type

KOmega <: AbstractTurbulenceModel

kOmega model containing all kOmega field parameters.

Fields

'k' – Turbulent kinetic energy ScalarField.

'omega' – Specific dissipation rate ScalarField.

'nut' – Eddy viscosity ScalarField.

'kf' – Turbulent kinetic energy FaceScalarField.

'omegaf' – Specific dissipation rate FaceScalarField.

'nutf' – Eddy viscosity FaceScalarField.

'coeffs' – Model coefficients.

XCALibre.ModelPhysics.KOmegaLKE — Type

KOmegaLKE <: AbstractTurbulenceModel

kOmega model containing all kOmega field parameters.

Fields

'k' – Turbulent kinetic energy ScalarField.

'omega' – Specific dissipation rate ScalarField.

'kl' – ScalarField.

'nut' – Eddy viscosity ScalarField.

'kf' – Turbulent kinetic energy FaceScalarField.

'omegaf' – Specific dissipation rate FaceScalarField.

'klf' – FaceScalarField.

'nutf' – Eddy viscosity FaceScalarField.

'coeffs' – Model coefficients.

'Tu' – Freestream turbulence intensity for model.

'y' – Near-wall distance for model.

XCALibre.ModelPhysics.LES — Type

LES <: AbstractLESModel

Abstract LES model type for consturcting LES models.

Fields

'args' – Model arguments.

XCALibre.ModelPhysics.Laminar — Type

Laminar <: AbstractTurbulenceModel

Laminar model definition for physcis API.

XCALibre.ModelPhysics.Momentum — Type

Momentum

Momentum model containting key momentum fields.

Fields

'U' – Velocity VectorField.

'p' – Pressure ScalarField.

'sources' – Momentum model sources.

Examples

`Momentum(mesh::AbstractMesh)

XCALibre.ModelPhysics.Physics — Type

Physics

XCALibre's physcis model API.

Fields

'time' – Time model.

'fluid' – Fluid model.

'momentum' – Momentum model.

'turbulence' – Turbulence model.

'energy' – Energy model.

'domain' – Mesh.

'boundary_info' – Mesh boundardy information.

Examples

`Phycsics( time::Union{Steady, Transient}, fluid::AbstractFluid, turbulence::AbstractTurbulenceModel, energy::AbstractEnergyModel, domain::AbstractMesh )

XCALibre.ModelPhysics.RANS — Type

RANS <: AbstractRANSModel

Abstract RANS model type for consturcting RANS models.

Fields

'args' – Model arguments.

XCALibre.ModelPhysics.SensibleEnthalpy — Type

SensibleEnthalpy <: AbstractEnergyModel

Type that represents energy model, coefficients and respective fields.

Fields

'h' – Sensible enthalpy ScalarField.

'T' – Temperature ScalarField.

'hf' – Sensible enthalpy FaceScalarField.

'Tf' – Temperature FaceScalarField.

'K' – Specific kinetic energy ScalarField.

'dpdt' – Pressure time derivative ScalarField.

'updated_BC' – Boundary condition function to convert temperature to sensible enthalp on on a fixed value boudary.

'coeffs' – A tuple of model coefficients.

XCALibre.ModelPhysics.Smagorinsky — Type

Smagorinsky <: AbstractTurbulenceModel

Smagorinsky LES model containing all Smagorinksy field parameters.

Fields

'nut' – Eddy viscosity ScalarField.

'nutf' – Eddy viscosity FaceScalarField.

'coeffs' – Model coefficients.

XCALibre.ModelPhysics.Steady — Type

Steady

Steady model for Physics model API.

Examples

Steady()

XCALibre.ModelPhysics.Transient — Type

Transient

Transient model for Physics model API.

Examples

Transient()

XCALibre.ModelPhysics.WeaklyCompressible — Type

WeaklyCompressible <: AbstractCompressible

Weakly compressible fluid model containing fluid field parameters for weakly compressible flows with constant parameters - ideal gas with constant viscosity.

Fields

'nu' – Fluid kinematic viscosity.

'cp' – Fluid specific heat capacity.

gamma – Ratio of specific heats.

Pr – Fluid Prantl number.

Examples

Fluid{WeaklyCompressible}(; nu=1E-5, cp=1005.0, gamma=1.4, Pr=0.7) - Constructor with

default values.

XCALibre.ModelPhysics.Ttoh! — Method

Ttoh!(model::Physics{T1,F,M,Tu,E,D,BI}, T::ScalarField, h::ScalarField
) where {T1,F<:AbstractCompressible,M,Tu,E,D,BI}

Function coverts temperature ScalarField to sensible enthalpy ScalarField.

Input

model – Physics model defined by user.

T – Temperature ScalarField.

h – Sensible enthalpy ScalarField.

XCALibre.ModelPhysics.energy! — Method

energy::Sensible_Enthalpy_Model{E1}, model::Physics{T1,F,M,Tu,E,D,BI}, prev, mdotf, rho, mueff, time, config
) where {T1,F,M,Tu,E,D,BI,E1}

Run energy transport equations.

Input

energy – Energy model.

model – Physics model defined by user.

prev – Previous energy cell values.

mdtof – Face mass flow.

rho – Density ScalarField.

mueff – Effective viscosity FaceScalarField.

time –

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

XCALibre.ModelPhysics.htoT! — Method

htoT!(model::Physics{T1,F,M,Tu,E,D,BI}, h::ScalarField, T::ScalarField
) where {T1,F<:AbstractCompressible,M,Tu,E,D,BI}

Function coverts sensible enthalpy ScalarField to temperature ScalarField.

Input

model – Physics model defined by user.

h – Sensible enthalpy ScalarField.

T – Temperature ScalarField.

XCALibre.ModelPhysics.initialise — Method

initialise(energy::SensibleEnthalpy, model::Physics{T1,F,M,Tu,E,D,BI}, mdotf, rho, peqn, config
) where {T1,F,M,Tu,E,D,BI})

Initialisation of energy transport equations.

Input

energy – Energy model.

model – Physics model defined by user.

mdtof – Face mass flow.

rho – Density ScalarField.

peqn – Pressure equation.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

Output

Sensible_Enthalpy_Model – Energy model struct containing energy equation.

XCALibre.ModelPhysics.initialise — Method

initialise(turbulence::KOmega, model::Physics{T,F,M,Tu,E,D,BI}, mdotf, peqn, config
) where {T,F,M,Tu,E,D,BI}

Initialisation of turbulent transport equations.

Input

turbulence – turbulence model.

model – Physics model defined by user.

mdtof – Face mass flow.

peqn – Pressure equation.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

Output

KOmegaModel(k_eqn, ω_eqn) – Turbulence model structure.

XCALibre.ModelPhysics.initialise — Method

initialise(turbulence::KOmegaLKE, model::Physics{T,F,M,Tu,E,D,BI}, mdotf, peqn, config
) where {T,F,M,Tu,E,D,BI}

Initialisation of turbulent transport equations.

Input

turbulence – turbulence model.

model – Physics model defined by user.

mdtof – Face mass flow.

peqn – Pressure equation.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

Output

KOmegaLKEModel( k_eqn, ω_eqn, kl_eqn, nueffkLS, nueffkS, nueffωS, nuL, nuts, Ω, γ, fv, normU, Reυ, ∇k, ∇ω ) – Turbulence model structure.

XCALibre.ModelPhysics.initialise — Method

turbulence::Laminar, model::Physics{T,F,M,Tu,E,D,BI}, mdotf, peqn, config
) where {T,F,M,Tu,E,D,BI}

Initialisation of turbulent transport equations.

Input

turbulence – turbulence model.

model – Physics model defined by user.

mdtof – Face mass flow.

peqn – Pressure equation.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

Output

LaminarModel() – Turbulence model structure.

XCALibre.ModelPhysics.initialise — Method

initialise(turbulence::Smagorinsky, model::Physics{T,F,M,Tu,E,D,BI}, mdotf, peqn, config
) where {T,F,M,Tu,E,D,BI}

Initialisation of turbulent transport equations.

Input

turbulence – turbulence model.

model – Physics model defined by user.

mdtof – Face mass flow.

peqn – Pressure equation.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

Output

SmagorinskyModel(Δ, magS) – Turbulence model structure.

XCALibre.ModelPhysics.thermo_Psi! — Method

thermo_Psi!(model::Physics{T,F,M,Tu,E,D,BI}, Psif::FaceScalarField) 
where {T,F<:AbstractCompressible,M,Tu,E,D,BI}

Function updates the value of Psi.

Input

model – Physics model defined by user.

Psif – Compressibility factor FaceScalarField.

Algorithm

Weakly compressible currently uses the ideal gas equation for establishing the compressibility factor where $\rho = p * \Psi$. $\Psi$ is calculated from the sensible enthalpy, reference temperature and fluid model specified $C_p$ and $R$ value where $R$ is calculated from $C_p$ and $\gamma$ specified in the fluid model.

XCALibre.ModelPhysics.thermo_Psi! — Method

thermo_Psi!(model::Physics{T,F,M,Tu,E,D,BI}, Psi::ScalarField) 
where {T,F<:AbstractCompressible,M,Tu,E,D,BI}

Model updates the value of Psi.

Input

model – Physics model defined by user.

Psi – Compressibility factor ScalarField.

Algorithm

Weakly compressible currently uses the ideal gas equation for establishing the compressibility factor where $\rho = p * \Psi$. $\Psi$ is calculated from the sensible enthalpy, reference temperature and fluid model specified $C_p$ and $R$ value where $R$ is calculated from $C_p$ and $\gamma$ specified in the fluid model.

XCALibre.ModelPhysics.turbulence! — Method

turbulence!(rans::KOmegaLKEModel, model::Physics{T,F,M,Turb,E,D,BI}, S, S2, prev, time, config ) where {T,F,M,Turb<:KOmegaLKE,E,D,BI}

Run turbulence model transport equations.

Input

rans::KOmegaLKEModel – KOmega turbulence model.

model – Physics model defined by user.

S – Strain rate tensor.

S2 – Square of the strain rate magnitude.

prev – Previous field.

time –

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

XCALibre.ModelPhysics.turbulence! — Method

turbulence!(rans::LaminarModel, model::Physics{T,F,M,Tu,E,D,BI}, S, S2, prev, time, config
) where {T,F,M,Tu<:Laminar,E,D,BI}

Run turbulence model transport equations.

Input

rans::LaminarModel – Laminar turbulence model.

model – Physics model defined by user.

S – Strain rate tensor.

S2 – Square of the strain rate magnitude.

prev – Previous field.

time –

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

XCALibre.ModelPhysics.turbulence! — Method

turbulence!(rans::KOmegaModel{E1,E2}, model::Physics{T,F,M,Tu,E,D,BI}, S, S2, prev, time, config
) where {T,F,M,Tu<:KOmega,E,D,BI,E1,E2}

Run turbulence model transport equations.

Input

rans::KOmegaModel{E1,E2} – KOmega turbulence model.

model – Physics model defined by user.

S – Strain rate tensor.

S2 – Square of the strain rate magnitude.

prev – Previous field.

time –

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

XCALibre.ModelPhysics.turbulence! — Method

turbulence!(les::SmagorinskyModel{E1,E2}, model::Physics{T,F,M,Tu,E,D,BI}, S, S2, prev, time, config
) where {T,F,M,Tu<:Smagorinsky,E,D,BI,E1,E2}

Run turbulence model transport equations.

Input

les::SmagorinskyModel{E1,E2} – Smagorinsky LES turbulence model.

model – Physics model defined by user.

S – Strain rate tensor.

S2 – Square of the strain rate magnitude.

prev – Previous field.

time –

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

XCALibre.Solvers.csimple! — Method

csimple!(model_in, config; resume=true, pref=nothing)

Compressible variant of the SIMPLE algorithm with a sensible enthalpy transport equation for the energy.

Input

model in – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

resume – True or false indicating if case is resuming or starting a new simulation.

pref – Reference pressure value for cases that do not have a pressure defining BC (incompressible flows only)

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.piso_comp! — Method

piso_comp!(model_in, config; resume=true, pref=nothing)

Compressible variant of the PISO algorithm with a sensible enthalpy transport equation for the energy.

Input

model in – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

resume – True or false indicating if case is resuming or starting a new simulation.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Steady,F<:Incompressible,M,Tu,E,D,BI}

Incompressible steady solver using the SIMPLE algorithm.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Steady,F<:WeaklyCompressible,M,Tu,E,D,BI}

Mildly compressible steady solver using the SIMPLE algorithm for low speed cases with heat transfer.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Transient,F<:Incompressible,M,Tu,E,D,BI}

Incompressible unsteady solver using the PISO algorithm.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Transient,F<:WeaklyCompressible,M,Tu,E,D,BI}

Mildly compressible unsteady solver using the PISO algorithm for low speed cases with heat transfer.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.simple! — Method

simple!(model_in, config; resume=true, pref=nothing)

Compressible variant of the SIMPLE algorithm with a sensible enthalpy transport equation for the energy.

Input

model in – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

resume – True or false indicating if case is resuming or starting a new simulation.

pref – Reference pressure value for cases that do not have a pressure defining BC (incompressible flows only)

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.