Runtime and solvers

Final steps before launching a simulation

Runtime setup

XCALibre.Solve.set_runtime — Function

set_runtime(; 
    # keyword arguments
    iterations::I, 
    write_interval::I, 
    time_step::N
    ) where {I<:Integer,N<:Number} = begin
    
    # returned `NamedTuple``
        (
        iterations=iterations, 
        dt=time_step, 
        write_interval=write_interval)
end

This is a convenience function to set the top-level runtime information. The inputs are all keyword arguments and provide basic information to flow solvers just before running a simulation.

Input arguments

iterations::Integer specifies the number of iterations in a simulation run.

write_interval::Integer define how often simulation results are written to file. The interval is currently based on number of iterations. Set to -1 to run without writing results to file.

time_step::Number the time step to use in the simulation. Notice that for steady solvers this is simply a counter and it is recommended to simply use 1.

Example

runtime = set_runtime(
    iterations=2000, time_step=1, write_interval=2000)

Configuration object

XCALibre.Simulate.Configuration — Type

@kwdef struct Configuration{SC,SL,RT,HW}
    schemes::SC
    solvers::SL
    runtime::RT
    hardware::HW
end

The Configuration type is pass to all flow solvers and provides all the relevant configuration to run a simulation.

Inputs

schemes::NamedTuple this keyword argument is used to pass distretisation scheme information to flow solvers. See Numerical setup for details.

solvers::NamedTuple this keyword argument is used to pass the configurations for the linear solvers for each field information to flow solvers. See Runtime and solvers for details.

runtime::NamedTuple this keyword argument is used to pass runtime information to the flow solvers. See Runtime and solvers for details.

hardware::NamedTuple this keyword argument is used to pass the hardware configuration and backend settings to the flow solvers. See Pre-processing for details.

Example

config = Configuration(
    solvers=solvers, schemes=schemes, runtime=runtime, hardware=hardware)

Initialising fields

XCALibre.Fields.initialise! — Function

function initialise!(field, value) # dummy function for documentation
    # Assign `value` to field in-place
    nothing
end

This function will set the given field to the value provided in-place. Useful for initialising fields prior to running a simulation.

Input arguments

field specifies the field to be initialised. The field must be either a AbractScalarField or AbstractVectorField

value defines the value to be set. This should be a scalar or vector (3 components) depending on the field to be modified e.g. for an AbstractVectorField we can specify as value=[10,0,0]

Note: in most cases the fields to be modified are stored within a physics model i.e. a Physics object. Thus, the argument value must fully qualify the model. For example, if we have created a Physics model named mymodel to set the velocity field, U, we would set the argument field to mymodel.momentum.U. See the example below.

Example

initialise!(mymodel.momentum.U, [2.5, 0, 0])
initialise!(mymodel.momentum.p, 1.25)

Launching flow solvers

XCALibre.Solvers.run! — Method

residuals = run!(model::Physics, config::Configuration)

Description goes here

Input arguments

Example

Something here