Numerical setup

Available discretisation schemes, linear solvers and preconditioners

Discretisation schemes

Part of the methodology to solve the various model equations using the Finite Volume Method (FVM) is to discretise each equation term, essentially, this process linearises the partial differential equation so that it can be represented as a system of linear equations, which can be solved using linear algebra along with iterative solvers. This section presents the discretisation schemes currently available in XCALibre.jl.

Discretisation schemes in XCALibre.jl are organised under the abstract type AbstractScheme. As shown previously, a list of available schemes can be found using the subtypes function:

AbstractScheme
├─ BoundedUpwind
├─ CrankNicolson
├─ Euler
├─ LUST
├─ Linear
├─ Midpoint
├─ Orthogonal
├─ SteadyState
└─ Upwind

Schemes available

Time schemes

Scheme	Description

SteadyState	sets the time derivative to zero

Euler	First order implicit Euler scheme

CrankNicolson	Second order central scheme (not implemented yet)

Laplacian schemes

Scheme	Description

Linear	2nd order Gauss gradient scheme with linear interpolation

Divergence schemes

Scheme	Description

Linear	2nd order central difference

Upwind	1st order upwind scheme

BoundedUpwind	Bounded version of the Upwind scheme

LUST	1st/2nd order mixed scheme (fixed at 75% Linear - 25% Upwind)

Gradient schemes

Scheme	Description

Orthogonal	Green-Gauss uncorrected gradient scheme

Midpoint	Green-Gauss skew corrected scheme (2 iterations - hardcoded)

Gradient limiters

Scheme	Description

FaceBased	Limit gradient calculation based on face extrapolation

MFaceBased	Limit gradient calculation based on face extrapolation in multiple dimensions

The construction of all gradient limiters currently requires users to pass the mesh object that will be used for the simulations e.g. MFaceBased(mesh). Notice that the mesh object must be moved to the same device where the simulation will be performed. To prevent mistakes and avoid errors, it is recommended to construct the limiters using the simulation Physics object directly since it provide a reference to the mesh (domain) e.g. if a Physics object named model has been created, construct the limiter as MFaceBased(model.domain)

Specifying schemes

XCALibre.jl flow solvers offer considerable flexibility to users for defining discretisation schemes. However, this means that discretisation schemes must be specified for every term of every equation solved. The schemes must be provided as a NamedTuple where each keyword corresponds to the fields being solved, e.g. (U = ..., p = ..., k = ..., <field> = ...). To facilitate this process, the set_schemes function is provided. Used without any inputs set_schemes uses the default values provided (see details below).

XCALibre.Discretise.set_schemes — Function

set_schemes(;
    # keyword arguments and their default values
    time=SteadyState,
    divergence=Linear, 
    laplacian=Linear, 
    gradient=Orthogonal,
    limiter=nothing) = begin
    
    # Returns NamedTuple definition for scheme 
    (
        time=time,
        divergence=divergence,
        laplacian=laplacian,
        gradient=gradient,
        limiter=limiter
    )   
end

The set_schemes function is used at the top-level API to help users define discretisation schemes for every field solved. It offers default values, thus users can pick and choose which entry they wish to modify.

inputs

time is used to set the time schemes (default is SteadyState)

divergence is used to set the divergence scheme (default is Linear)

laplacian is used to set the laplacian scheme (default is Linear)

gradient is used to set the gradient scheme (default is Orthogonal)

limiter is used to specify if gradient limiters should be used, currently supported limiters include FaceBased and MFaceBased (default is nothing)

For example, below we set the schemes for the U and p fields. Notice that in the first case the schemes will take their default values (entry for p). In the case of U, we are only changing the setting for the divergence scheme to Upwind.

using XCALibre

# Note: this example assumes a Physics object named `model` already exists

schemes = (
    p = set_schemes(), # no input provided (will use defaults)
    U = set_schemes(divergence = Upwind, gradient=Midpoint, limiter=MFaceBased(model.domain)),
)

Linear solvers

Linear solvers in XCALibre.jl are provided by Krylov.jl. The following solvers types are re-exported in XCALibre.jl

BicgstabSolver is a general purpose linear solver. Works well with non-symmetric matrices e.g. for U.

CgSolver is particular strong with symmetric matrices e.g to solve the pressure equation.

GmresSolver is a general solver. We have found it works best on the CPU backend.

For more information on these solvers you can review the excellent documentation provided by the Krylov.jl team.

XCALibre.jl provides the set_solver convenience function for setting solvers. See details below.

XCALibre.Solve.set_solver — Function

set_solver( 
    field::AbstractField;
    # keyword arguments and defaults
    solver::S, 
    preconditioner::PT, 
    convergence, 
    relax,
    smoother=nothing,
    limit=(),
    itmax::Integer=1000, 
    atol=(eps(_get_float(field.mesh)))^0.9,
    rtol=_get_float(field.mesh)(1e-1)
    ) where {S,PT<:PreconditionerType} = begin

    # return NamedTuple
    TF = _get_float(field.mesh)
    (
        solver=solver, 
        preconditioner=preconditioner, 
        convergence=convergence |> TF, 
        relax=relax |> TF, 
        smoother=smoother,
        limit=limit,
        itmax=itmax, 
        atol=atol |> TF, 
        rtol=rtol |> TF
    )
end

This function is used to provide solver settings that will be used internally in XCALibre.jl. It returns a NamedTuple with solver settings that are used internally by the flow solvers.

Input arguments

field reference to the field to which the solver settings will apply (used to provide integer and float types required)

solver solver object from Krylov.jl and it could be one of BicgstabSolver, CgSolver, GmresSolver which are re-exported in XCALibre.jl

preconditioner instance of preconditioner to be used e.g. Jacobi()

convergence sets the stopping criteria of this field

relax specifies the relaxation factor to be used e.g. set to 1 for no relaxation

smoother specifies smoothing method to be applied before discretisation. JacobiSmoother is currently the only choice (defaults to nothing)

limit used in some solvers to bound the solution within this limits e.g. (min, max). It defaults to ()

itmax maximum number of iterations in a single solver pass (defaults to 1000)

atol absolute tolerance for the solver (default to eps(FloatType)^0.9)

rtol set relative tolerance for the solver (defaults to 1e-1)

Preconditioners

XCALibre.jl offers a range of preconditioners which are subtypes of the abstract type PreconditionerType, exploring its subtypes we can find a list of the currently available preconditioners:

PreconditionerType
├─ LDIVPreconditioner
│  └─ DILU
└─ MULPreconditioner
   ├─ IC0GPU
   ├─ ILU0GPU
   ├─ Jacobi
   └─ NormDiagonal

Note

Only the Jacobi and NormDiagonal preconditioners have GPU ready implementations. At present these have the most robust implementation and they can be used with both CPU and GPU backends. The other preconditioners can only be used on the CPU. Notice that on our tests the LDL preconditioner only works when paired with the GmresSolver on the CPU. Also notice that the implementations for DILU (experimental), IC0GPU and ILU0GPU (NVIDIA only), should be considered experimental. Work on improving the offering of preconditioners is ongoing.

Warning

Internally the storage for sparse matrices was moved to the CSR format. Thus, temporarily, we have removed support for LDL and ILU0 preconditioners while we work on CSR compatible implementations.

Below an example is provided in context. Here, we are setting solvers for both the velocity field U and the pressure field p and packing them into a NamedTuple "solvers". The Jacobi preconditioner is used in both solvers. Notice that preconditioners are specified with an instance of their type i.e. Jacobi(). Internally, the preconditioner instance is used for dispatch. This tupple will then be passed on to create the final Configuration object.

using XCALibre

# Note: this example assumes a Physics object named `model` already exists

solvers = (
    U = set_solver(
        model.momentum.U;
        solver      = BicgstabSolver, # GmresSolver
        preconditioner = Jacobi(),
        convergence = 1e-7,
        relax       = 0.7,
        rtol = 1e-4,
        atol = 1e-10
    ),
    p = set_solver(
        model.momentum.p;
        solver      = CgSolver, # BicgstabSolver, GmresSolver
        preconditioner = Jacobi(),
        convergence = 1e-7,
        relax       = 0.7,
        rtol = 1e-4,
        atol = 1e-10
    )
)