Reference

XCALibre.ModelPhysics.initialise — Method

initialise(energy::SensibleEnthalpy, model::Physics{T1,F,M,Tu,E,D,BI}, mdotf, rho, peqn, config
) where {T1,F,M,Tu,E,D,BI})

Initialisation of energy transport equations.

Input

energy – Energy model.

model – Physics model defined by user.

mdtof – Face mass flow.

rho – Density ScalarField.

peqn – Pressure equation.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

Output

`SensibleEnthalpyModel' – Energy model struct containing energy equation.

XCALibre.Solvers.csimple! — Method

csimple!(model_in, config; resume=true, pref=nothing)

Compressible variant of the SIMPLE algorithm with a sensible enthalpy transport equation for the energy.

Input

model in – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

resume – True or false indicating if case is resuming or starting a new simulation.

pref – Reference pressure value for cases that do not have a pressure defining BC (incompressible flows only)

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.piso_comp! — Method

piso_comp!(model_in, config; resume=true, pref=nothing)

Compressible variant of the PISO algorithm with a sensible enthalpy transport equation for the energy.

Input

model in – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

resume – True or false indicating if case is resuming or starting a new simulation.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Steady,F<:Incompressible,M,Tu,E,D,BI}

Incompressible steady solver using the SIMPLE algorithm.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Steady,F<:WeaklyCompressible,M,Tu,E,D,BI}

Mildly compressible steady solver using the SIMPLE algorithm for low speed cases with heat transfer.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Transient,F<:Incompressible,M,Tu,E,D,BI}

Incompressible unsteady solver using the PISO algorithm.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.run! — Method

run!(
model::Physics{T,F,M,Tu,E,D,BI}, config; pref=nothing
) where{T<:Transient,F<:WeaklyCompressible,M,Tu,E,D,BI}

Mildly compressible unsteady solver using the PISO algorithm for low speed cases with heat transfer.

Input

model – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

pref – Reference pressure value for cases that do not have a pressure defining BC.

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.

XCALibre.Solvers.simple! — Method

simple!(model_in, config; resume=true, pref=nothing)

Compressible variant of the SIMPLE algorithm with a sensible enthalpy transport equation for the energy.

Input

model in – Physics model defiend by user and passed to run!.

config – Configuration structure defined by user with solvers, schemes, runtime and hardware structures set.

resume – True or false indicating if case is resuming or starting a new simulation.

pref – Reference pressure value for cases that do not have a pressure defining BC (incompressible flows only)

Output

R_ux - Vector of x-velocity residuals for each iteration.

R_uy - Vector of y-velocity residuals for each iteration.

R_uz - Vector of y-velocity residuals for each iteration.

R_p - Vector of pressure residuals for each iteration.

R_e - Vector of energy residuals for each iteration.

model - Physics model output including field parameters.